+++ axelor-account/src/main/java/com/axelor/apps/account/service/invoice/print/InvoicePrintServiceImpl.java
@@ -33,7 +33,6 @@ import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
 import com.axelor.meta.MetaFiles;
 import com.axelor.meta.db.MetaFile;
-import com.google.common.collect.Iterables;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.google.inject.persist.Transactional;
@@ -49,7 +48,7 @@ import java.util.List;
 public class InvoicePrintServiceImpl implements InvoicePrintService {
 
   @Inject private InvoiceRepository invoiceRepo;
-  
+
   @Override
   public String printInvoice(Invoice invoice, boolean forceRefresh)
       throws AxelorException, IOException {
@@ -119,14 +118,15 @@ public class InvoicePrintServiceImpl implements InvoicePrintService {
   @Override
   public String printInvoices(List<Long> ids) throws IOException, AxelorException {
     List<File> printedInvoices = new ArrayList<>();
-    List<String> invalidPrintSettingsInvoiceIds=checkInvalidPrintSettingsInvoices(ids);
-    
-    if(invalidPrintSettingsInvoiceIds.size() > 0){
+    List<String> invalidPrintSettingsInvoiceIds = checkInvalidPrintSettingsInvoices(ids);
+
+    if (invalidPrintSettingsInvoiceIds.size() > 0) {
       throw new AxelorException(
-          TraceBackRepository.CATEGORY_MISSING_FIELD,I18n.get(IExceptionMessage.INVOICES_MISSING_PRINTING_SETTINGS),
+          TraceBackRepository.CATEGORY_MISSING_FIELD,
+          I18n.get(IExceptionMessage.INVOICES_MISSING_PRINTING_SETTINGS),
           invalidPrintSettingsInvoiceIds.toString());
     }
-    
+
     ModelTool.apply(
         Invoice.class,
         ids,
@@ -135,19 +135,19 @@ public class InvoicePrintServiceImpl implements InvoicePrintService {
           public void accept(Invoice invoice) throws Exception {
             printedInvoices.add(printCopiesToFile(invoice, false));
           }
-        }); 
-       
+        });
+
     String fileName = getInvoiceFilesName(true);
     return PdfTool.mergePdfToFileLink(printedInvoices, fileName);
   }
-  
-public List<String>  checkInvalidPrintSettingsInvoices(List<Long> ids) {
- 
-  List<String> invalidPrintSettingsInvoiceIds=new ArrayList<>();
-  
-    for(Long id:ids) {
-      Invoice invoice=invoiceRepo.find(id);
-      if(invoice.getPrintingSettings() == null) {
+
+  public List<String> checkInvalidPrintSettingsInvoices(List<Long> ids) {
+
+    List<String> invalidPrintSettingsInvoiceIds = new ArrayList<>();
+
+    for (Long id : ids) {
+      Invoice invoice = invoiceRepo.find(id);
+      if (invoice.getPrintingSettings() == null) {
         invalidPrintSettingsInvoiceIds.add(invoice.getInvoiceId());
       }
     }